requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
et-xmlfile==1.1.0
flask==3.0.2
googlemaps==4.10.0
gunicorn==21.2.0
idna==3.6
importlib-metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.24.4
openpyxl==3.1.2
packaging==24.0
pandas==2.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
tzdata==2024.1
urllib3==2.2.1
werkzeug==3.0.1
zipp==3.18.1
